Refactor capability retrieval as a hook
