+++ workbench-app/src/routes/Interact.tsx
@@ -11,11 +11,10 @@ import { ConversationRename } from '../components/Conversations/ConversationRena
 import { ConversationShare } from '../components/Conversations/ConversationShare';
 import { InteractHistory } from '../components/Conversations/InteractHistory';
 import { InteractInput } from '../components/Conversations/InteractInput';
+import { useGetAssistantCapabilitiesSet } from '../libs/useAssistantCapabilities';
 import { useLocalUserAccount } from '../libs/useLocalUserAccount';
 import { useSiteUtility } from '../libs/useSiteUtility';
-import { useWorkbenchService } from '../libs/useWorkbenchService';
 import { Assistant } from '../models/Assistant';
-import { AssistantCapability } from '../models/AssistantCapability';
 import {
     useGetAssistantsInConversationQuery,
     useGetConversationFilesQuery,
@@ -101,10 +100,7 @@ export const Interact: React.FC = () => {
     } = useGetConversationFilesQuery(conversationId);
     const [updateConversation] = useUpdateConversationMutation();
     const { getUserId } = useLocalUserAccount();
-    const [assistantCapabilities, setAssistantCapabilities] = React.useState<Set<AssistantCapability> | undefined>(
-        undefined,
-    );
-    const workbenchService = useWorkbenchService();
+    const assistantCapabilities = useGetAssistantCapabilitiesSet(assistants ?? []);
 
     const siteUtility = useSiteUtility();
 
@@ -148,85 +144,6 @@ export const Interact: React.FC = () => {
             siteUtility.setDocumentTitle(conversation.title);
         }
     }, [conversation, conversationParticipants, siteUtility]);
-
-    // Build a memo-ized set of all capabilities to be used as a default for assistants that do not
-    // specify capabilities
-    const allCapabilities = React.useMemo(
-        () =>
-            Object.entries(AssistantCapability).reduce((acc, [_, capability]) => {
-                acc.add(capability);
-                return acc;
-            }, new Set<AssistantCapability>()),
-        [],
-    );
-
-    // Load the capabilities for all assistants and update the state with the result
-    React.useEffect(() => {
-        if (isLoadingAssistants || !assistants) {
-            return;
-        }
-
-        let ignore = false;
-
-        loadCapabilities();
-
-        return () => {
-            ignore = true;
-        };
-
-        async function loadCapabilities() {
-            if (isLoadingAssistants || !assistants) {
-                return;
-            }
-
-            if (assistants.length === 0) {
-                if (ignore) {
-                    return;
-                }
-                // only update the state if the capabilities have changed
-                if (!assistantCapabilities || allCapabilities.symmetricDifference(assistantCapabilities).size > 0) {
-                    setAssistantCapabilities(allCapabilities);
-                }
-                return;
-            }
-            // Get the service info for each assistant
-            const serviceInfos = (
-                await Promise.all(
-                    assistants.map(async (assistant) => {
-                        return await workbenchService.getAssistantServiceInfoAsync(assistant.assistantServiceId);
-                    }),
-                )
-            ).filter((info) => info !== undefined);
-
-            // Combine all capabilities from all assistants into a single set
-            const capabilities = serviceInfos.reduce<Set<AssistantCapability>>((acc, info) => {
-                const metadataCapabilities = info.metadata?.capabilities;
-
-                // If there are no capabilities specified at all, default to all capabilities
-                if (metadataCapabilities === undefined || Object.keys(metadataCapabilities).length === 0) {
-                    acc.union(allCapabilities);
-                    return acc;
-                }
-
-                const capabilitiesSet = new Set(
-                    Object.keys(metadataCapabilities)
-                        .filter((key) => metadataCapabilities[key])
-                        .map((key) => key as AssistantCapability),
-                );
-                acc = acc.union(capabilitiesSet);
-                return acc;
-            }, new Set<AssistantCapability>());
-
-            if (ignore) {
-                return;
-            }
-
-            // only update the state if the capabilities have changed
-            if (!assistantCapabilities || capabilities.symmetricDifference(assistantCapabilities).size > 0) {
-                setAssistantCapabilities(capabilities);
-            }
-        }
-    }, [allCapabilities, assistantCapabilities, assistants, isLoadingAssistants, workbenchService]);
 
     const handleConversationRename = React.useCallback(
         async (id: string, newTitle: string) => {
